start remaking articles using livewire and policies

+++ app/Http/Livewire/News/NewsSingle.php
@@ -2,16 +2,21 @@
 
 namespace App\Http\Livewire\News;
 
+use App\Models\News;
 use Livewire\Component;
 
 class NewsSingle extends Component
 {
-    public $news;
+    public News $news;
 
     public function deleteNews()
     {
+        if (auth()->user()->cannot('delete', $this->news)) {
+            abort(403);
+        }
+
         $this->news->delete();
-        $this->emit('news.deleted', $this->id);
+        $this->emit('news.deleted');
     }
 
     public function render()
